requirements.txt
playwright
python-dotenv
chromadb==0.4.6
langchain==0.0.300
google-generativeai
openai<1.0
fastapi==0.99.1
uvicorn
torch
numpy
pydantic<2.0
transformers==4.30.2
sentence-transformers==2.2.2